# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryRequest:
    properties:
      image_url:
        example: https://example.com/cat.jpg
        type: string
      name:
        example: Kaos Polos
        type: string
    required:
    - name
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryResponse:
    properties:
      created_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      image_url:
        example: https://example.com/cat.jpg
        type: string
      name:
        example: Kaos Polos
        type: string
      updated_at:
        example: "2023-10-01T15:00:00Z"
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_domain.PaginationMeta:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_faridlan_sikon-api_internal_utils.EmptyObj:
    type: object
  github_com_faridlan_sikon-api_internal_utils.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        example: pesan error
        type: string
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.PaginatedResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_CategoryResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryResponse'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_domain.PaginationMeta'
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_CategoryResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryResponse'
      message:
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj:
    properties:
      data:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.EmptyObj'
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Ini adalah dokumentasi API untuk MVP ERP SIKOn.
  title: SIKOn API (Sistem Integrasi Konveksi Online)
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Mengambil daftar seluruh kategori produk dengan pagination
      parameters:
      - default: 1
        description: Nomor Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Batas Data per Halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.PaginatedResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_CategoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Membuat kategori produk baru
      parameters:
      - description: Data Kategori Baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Menghapus kategori secara permanen
      parameters:
      - description: Category ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Category
      tags:
      - Categories
    get:
      description: Mengambil detail kategori berdasarkan ID
      parameters:
      - description: Category ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Memperbarui nama kategori
      parameters:
      - description: Category ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data Update Kategori
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Category
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"